+++ include/tao/pegtl/internal/file_mapper_posix.hpp
@@ -4,22 +4,74 @@
 #ifndef TAO_PEGTL_INTERNAL_FILE_MAPPER_POSIX_HPP
 #define TAO_PEGTL_INTERNAL_FILE_MAPPER_POSIX_HPP
 
+#include <fcntl.h>
 #include <sys/mman.h>
+#include <sys/stat.h>
+#include <sys/types.h>
 #include <unistd.h>
 
-#include <system_error>
+#include <filesystem>
+#include <utility>
 
 #include "../config.hpp"
 
-#include "file_opener.hpp"
-
 namespace TAO_PEGTL_NAMESPACE::internal
 {
+   struct file_opener
+   {
+      explicit file_opener( const std::filesystem::path& path )
+         : m_path( path ),
+           m_fd( open() )
+      {}
+
+      file_opener( const file_opener& ) = delete;
+      file_opener( file_opener&& ) = delete;
+
+      ~file_opener() noexcept
+      {
+         ::close( m_fd );
+      }
+
+      void operator=( const file_opener& ) = delete;
+      void operator=( file_opener&& ) = delete;
+
+      [[nodiscard]] std::size_t size() const
+      {
+         struct stat st;
+         errno = 0;
+         if( ::fstat( m_fd, &st ) < 0 ) {
+            const std::error_code ec( errno, std::system_category() );
+            throw std::filesystem::filesystem_error( "fstat() failed", m_path, ec );
+         }
+         return std::size_t( st.st_size );
+      }
+
+      const std::filesystem::path m_path;
+      const int m_fd;
+
+   private:
+      [[nodiscard]] int open() const
+      {
+         errno = 0;
+         const int fd = ::open( m_path.c_str(),
+                                O_RDONLY
+#if defined( O_CLOEXEC )
+                                   | O_CLOEXEC
+#endif
+         );
+         if( fd >= 0 ) {
+            return fd;
+         }
+         const std::error_code ec( errno, std::system_category() );
+         throw std::filesystem::filesystem_error( "open() failed", m_path, ec );
+      }
+   };
+
    class file_mapper
    {
    public:
-      explicit file_mapper( const char* filename )
-         : file_mapper( file_opener( filename ) )
+      explicit file_mapper( const std::filesystem::path& path )
+         : file_mapper( file_opener( path ) )
       {}
 
       explicit file_mapper( const file_opener& reader )
@@ -27,8 +79,8 @@ namespace TAO_PEGTL_NAMESPACE::internal
            m_data( static_cast< const char* >( ::mmap( nullptr, m_size, PROT_READ, MAP_PRIVATE, reader.m_fd, 0 ) ) )
       {
          if( ( m_size != 0 ) && ( intptr_t( m_data ) == -1 ) ) {
-            const auto ec = errno;
-            throw std::system_error( ec, std::system_category(), reader.m_source );
+            const std::error_code ec( errno, std::system_category() );
+            throw std::filesystem::filesystem_error( "mmap() failed", reader.m_path, ec );
          }
       }