+++ include/tao/pegtl/mmap_input.hpp
@@ -4,8 +4,7 @@
 #ifndef TAO_PEGTL_MMAP_INPUT_HPP
 #define TAO_PEGTL_MMAP_INPUT_HPP
 
-#include <string>
-#include <utility>
+#include <filesystem>
 
 #include "config.hpp"
 #include "eol.hpp"
@@ -29,13 +28,10 @@ namespace TAO_PEGTL_NAMESPACE
    {
       struct mmap_holder
       {
-         const std::string filename;
          const file_mapper data;
 
-         template< typename T >
-         explicit mmap_holder( T&& in_filename )
-            : filename( std::forward< T >( in_filename ) ),
-              data( filename.c_str() )
+         explicit mmap_holder( const std::filesystem::path& path )
+            : data( path )
          {}
 
          mmap_holder( const mmap_holder& ) = delete;
@@ -52,12 +48,11 @@ namespace TAO_PEGTL_NAMESPACE
    template< tracking_mode P = tracking_mode::eager, typename Eol = eol::lf_crlf >
    struct mmap_input
       : private internal::mmap_holder,
-        public memory_input< P, Eol, const char* >
+        public memory_input< P, Eol >
    {
-      template< typename T >
-      explicit mmap_input( T&& in_filename )
-         : internal::mmap_holder( std::forward< T >( in_filename ) ),
-           memory_input< P, Eol, const char* >( data.begin(), data.end(), filename.c_str() )
+      explicit mmap_input( const std::filesystem::path& path )
+         : internal::mmap_holder( path ),
+           memory_input< P, Eol >( data.begin(), data.end(), path.u8string() )
       {}
 
       mmap_input( const mmap_input& ) = delete;