+++ include/tao/pegtl/read_input.hpp
@@ -4,7 +4,7 @@
 #ifndef TAO_PEGTL_READ_INPUT_HPP
 #define TAO_PEGTL_READ_INPUT_HPP
 
-#include <string>
+#include <filesystem>
 
 #include "config.hpp"
 #include "eol.hpp"
@@ -15,43 +15,16 @@
 
 namespace TAO_PEGTL_NAMESPACE
 {
-   namespace internal
-   {
-      struct filename_holder
-      {
-         const std::string filename;
-
-         template< typename T >
-         explicit filename_holder( T&& in_filename )
-            : filename( std::forward< T >( in_filename ) )
-         {}
-
-         filename_holder( const filename_holder& ) = delete;
-         filename_holder( filename_holder&& ) = delete;
-
-         ~filename_holder() = default;
-
-         void operator=( const filename_holder& ) = delete;
-         void operator=( filename_holder&& ) = delete;
-      };
-
-   }  // namespace internal
-
    template< tracking_mode P = tracking_mode::eager, typename Eol = eol::lf_crlf >
    struct read_input
-      : private internal::filename_holder,
-        public string_input< P, Eol, const char* >
+      : string_input< P, Eol >
    {
-      template< typename T >
-      explicit read_input( T&& in_filename )
-         : internal::filename_holder( std::forward< T >( in_filename ) ),
-           string_input< P, Eol, const char* >( internal::file_reader( filename.c_str() ).read(), filename.c_str() )
+      explicit read_input( const std::filesystem::path& path )
+         : string_input< P, Eol >( internal::file_reader( path ).read(), path.u8string() )
       {}
 
-      template< typename T >
-      read_input( FILE* in_file, T&& in_filename )
-         : internal::filename_holder( std::forward< T >( in_filename ) ),
-           string_input< P, Eol, const char* >( internal::file_reader( in_file, filename.c_str() ).read(), filename.c_str() )
+      read_input( FILE* file, const std::filesystem::path& path )
+         : string_input< P, Eol >( internal::file_reader( file, path ).read(), path.u8string() )
       {}
 
       read_input( const read_input& ) = delete;